integrated_system/requirements.txt
# Q&ACE Integrated System - Requirements
# Combined requirements for multimodal emotion detection

# ============================================
# CORE DEEP LEARNING
# ============================================
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# ============================================
# COMPUTER VISION
# ============================================
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
facenet-pytorch>=2.5.0
Pillow>=10.0.0

# ============================================
# AUDIO PROCESSING
# ============================================
librosa>=0.10.0
soundfile>=0.12.0
sounddevice>=0.4.6
audioread>=3.0.0

# ============================================
# SPEECH-TO-TEXT (Whisper)
# ============================================
openai-whisper>=20231117

# ============================================
# DATA SCIENCE & ML
# ============================================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ============================================
# VISUALIZATION
# ============================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================
# WEB API
# ============================================
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# ============================================
# UTILITIES
# ============================================
tqdm>=4.65.0
requests>=2.31.0
safetensors>=0.4.0
python-dotenv>=1.0.0

# ============================================
# COACHING PIPELINE (Sentence-BERT)
# ============================================
sentence-transformers>=2.2.0

# ============================================
# GENAI (Gemini API)
# ============================================
google-generativeai>=0.3.0

# ============================================
# BERTSCORE (Semantic Scoring)
# ============================================
bert-score>=0.3.13
